Update base.py
The function of 'load_csv_with_header()' in this file is not flexible for user due to that user should define the number of features and samples at the first row. I re-write this sub function so that the users could directly use this function without define the number of features and number of a csv data.